chore: refactor to wrap client context in BigtableClientContext
