+++ google-cloud-bigtable/src/main/java/com/google/cloud/bigtable/data/v2/BigtableDataClientFactory.java
@@ -18,11 +18,9 @@ package com.google.cloud.bigtable.data.v2;
 import com.google.api.core.BetaApi;
 import com.google.api.gax.core.BackgroundResource;
 import com.google.api.gax.rpc.ClientContext;
+import com.google.cloud.bigtable.data.v2.stub.BigtableClientContext;
 import com.google.cloud.bigtable.data.v2.stub.EnhancedBigtableStub;
-import io.opentelemetry.api.OpenTelemetry;
 import java.io.IOException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import javax.annotation.Nonnull;
 
 /**
@@ -66,11 +64,8 @@ import javax.annotation.Nonnull;
 @BetaApi("This feature is currently experimental and can change in the future")
 public final class BigtableDataClientFactory implements AutoCloseable {
 
-  private static final Logger logger = Logger.getLogger(BigtableDataClientFactory.class.getName());
-
   private final BigtableDataSettings defaultSettings;
-  private final ClientContext sharedClientContext;
-  private final OpenTelemetry openTelemetry;
+  private final BigtableClientContext sharedClientContext;
 
   /**
    * Create a instance of this factory.
@@ -80,31 +75,16 @@ public final class BigtableDataClientFactory implements AutoCloseable {
    */
   public static BigtableDataClientFactory create(BigtableDataSettings defaultSettings)
       throws IOException {
-    ClientContext sharedClientContext =
-        EnhancedBigtableStub.createClientContext(defaultSettings.getStubSettings());
-    OpenTelemetry openTelemetry = null;
-    try {
-      // We don't want client side metrics to crash the client, so catch any exception when getting
-      // the OTEL instance and log the exception instead.
-      openTelemetry =
-          EnhancedBigtableStub.getOpenTelemetry(
-              defaultSettings.getProjectId(),
-              defaultSettings.getMetricsProvider(),
-              sharedClientContext.getCredentials(),
-              defaultSettings.getStubSettings().getMetricsEndpoint());
-    } catch (Throwable t) {
-      logger.log(Level.WARNING, "Failed to get OTEL, will skip exporting client side metrics", t);
-    }
-    return new BigtableDataClientFactory(sharedClientContext, defaultSettings, openTelemetry);
+    BigtableClientContext sharedClientContext =
+        EnhancedBigtableStub.createBigtableClientContext(defaultSettings.getStubSettings());
+
+    return new BigtableDataClientFactory(sharedClientContext, defaultSettings);
   }
 
   private BigtableDataClientFactory(
-      ClientContext sharedClientContext,
-      BigtableDataSettings defaultSettings,
-      OpenTelemetry openTelemetry) {
+      BigtableClientContext sharedClientContext, BigtableDataSettings defaultSettings) {
     this.sharedClientContext = sharedClientContext;
     this.defaultSettings = defaultSettings;
-    this.openTelemetry = openTelemetry;
   }
 
   /**
@@ -114,7 +94,8 @@ public final class BigtableDataClientFactory implements AutoCloseable {
    */
   @Override
   public void close() throws Exception {
-    for (BackgroundResource resource : sharedClientContext.getBackgroundResources()) {
+    for (BackgroundResource resource :
+        sharedClientContext.getClientContext().getBackgroundResources()) {
       resource.close();
     }
   }
@@ -132,10 +113,11 @@ public final class BigtableDataClientFactory implements AutoCloseable {
     try {
       ClientContext clientContext =
           sharedClientContext
+              .getClientContext()
               .toBuilder()
               .setTracerFactory(
                   EnhancedBigtableStub.createBigtableTracerFactory(
-                      defaultSettings.getStubSettings(), openTelemetry))
+                      defaultSettings.getStubSettings(), sharedClientContext.createOpenTelemetry()))
               .build();
 
       return BigtableDataClient.createWithClientContext(defaultSettings, clientContext);
@@ -161,10 +143,11 @@ public final class BigtableDataClientFactory implements AutoCloseable {
 
     ClientContext clientContext =
         sharedClientContext
+            .getClientContext()
             .toBuilder()
             .setTracerFactory(
                 EnhancedBigtableStub.createBigtableTracerFactory(
-                    settings.getStubSettings(), openTelemetry))
+                    settings.getStubSettings(), sharedClientContext.createOpenTelemetry()))
             .build();
     return BigtableDataClient.createWithClientContext(settings, clientContext);
   }
@@ -190,10 +173,11 @@ public final class BigtableDataClientFactory implements AutoCloseable {
 
     ClientContext clientContext =
         sharedClientContext
+            .getClientContext()
             .toBuilder()
             .setTracerFactory(
                 EnhancedBigtableStub.createBigtableTracerFactory(
-                    settings.getStubSettings(), openTelemetry))
+                    settings.getStubSettings(), sharedClientContext.createOpenTelemetry()))
             .build();
 
     return BigtableDataClient.createWithClientContext(settings, clientContext);
@@ -220,10 +204,11 @@ public final class BigtableDataClientFactory implements AutoCloseable {
             .build();
     ClientContext clientContext =
         sharedClientContext
+            .getClientContext()
             .toBuilder()
             .setTracerFactory(
                 EnhancedBigtableStub.createBigtableTracerFactory(
-                    settings.getStubSettings(), openTelemetry))
+                    settings.getStubSettings(), sharedClientContext.createOpenTelemetry()))
             .build();
     return BigtableDataClient.createWithClientContext(settings, clientContext);
   }
